Moved to 64bit transforms, the camera and multiple other stuff probably still use 32bit, needs moving later, intention is to make cpu side 64bit.

+++ Engine/Source/Systems/ECS/Components/TransformComponent.h
@@ -6,17 +6,17 @@
 namespace Aozora {
 
 	struct TransformComponent {
-		glm::mat4 model = glm::mat4(1.0f); // mat4 storage of the pos/rot/scale
-		glm::mat4 baseModel = glm::mat4(1.0f);
-		glm::vec3 pos = glm::vec3(0.0f);
-		glm::vec3 rot = glm::vec3(0.0f);
-		glm::vec3 scale = glm::vec3(1.0f);
+		glm::dmat4 model = glm::dmat4(1.0f); // mat4 storage of the pos/rot/scale
+		glm::dmat4 baseModel = glm::dmat4(1.0f);
+		glm::dvec3 pos = glm::dvec3(0.0f);
+		glm::dvec3 rot = glm::dvec3(0.0f);
+		glm::dvec3 scale = glm::dvec3(1.0f);
 
 		bool isDirty{ true };
 
 		template<class Archive>
 		void serialize(Archive& archive) {
-			archive(CEREAL_NVP(model),
+			archive(CEREAL_NVP(model),	
 				CEREAL_NVP(baseModel),
 				CEREAL_NVP(pos),
 				CEREAL_NVP(rot),
